% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{as.coefficients_textmodel}
\alias{as.coefficients_textmodel}
\title{Coerce various objects to coefficients_textmodel}
\usage{
as.coefficients_textmodel(x)
}
\arguments{
\item{x}{an object to be coerced}
}
\value{
an object with the class tag of \code{coefficients_textmodel}
}
\description{
Helper functions used in \verb{summary.textmodel_*()}.
}
\keyword{internal}
