% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_svm.R
\name{as.matrix.csr.dfm}
\alias{as.matrix.csr.dfm}
\title{convert a dfm into a matrix.csr from SparseM package}
\usage{
\method{as.matrix.csr}{dfm}(x, ...)
}
\arguments{
\item{x}{input \link{dfm}}

\item{...}{unused}
}
\value{
a \pkg{SparseM} object of class \link[SparseM:SparseM.ontology]{matrix.csr}
}
\description{
Utility to convert a dfm into a \link[SparseM:SparseM.ontology]{matrix.csr} from the \pkg{SparseM} package.
}
\keyword{internal}
