% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel-methods.R
\name{as.summary.textmodel}
\alias{as.summary.textmodel}
\title{Assign the summary.textmodel class to a list}
\usage{
as.summary.textmodel(x)
}
\arguments{
\item{x}{a named list}
}
\value{
an object of class \code{summary.textmodel}
}
\description{
Assigns the class \code{summary.textmodel} to a list
}
\keyword{internal}
