% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens2sequences.R
\name{is.tokens2sequences}
\alias{is.tokens2sequences}
\title{Check to see if function is a tokens2sequences type}
\usage{
is.tokens2sequences(x)
}
\arguments{
\item{x}{Object that will be checked to see if it is of the type \code{\link{tokens2sequences}}}
}
\description{
Check to see if function is a tokens2sequences type
}
\seealso{
\code{\link{tokens2sequences}}
}
\keyword{internal}
