% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_cnnlstmemb.R
\name{load.textmodel_cnnlstmemb}
\alias{load.textmodel_cnnlstmemb}
\title{A function for loading saved keras-based models}
\usage{
\method{load}{textmodel_cnnlstmemb}(x, ...)
}
\arguments{
\item{x}{saved model generated by \code{\link{textmodel_cnnlstmemb}}}

\item{...}{additional arguments passed to load}
}
\description{
A function for loading saved keras-based models
}
