% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_nnseq.R
\name{load.textmodel_nnseq}
\alias{load.textmodel_nnseq}
\title{A function for loading saved keras-based models}
\usage{
\method{load}{textmodel_nnseq}(x, ...)
}
\arguments{
\item{x}{saved model generated by \code{\link{textmodel_nnseq}}}

\item{...}{additional arguments passed to load}
}
\description{
A function for loading saved keras-based models
}
