% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_nnseq.R
\name{predict.textmodel_nnseq}
\alias{predict.textmodel_nnseq}
\title{Prediction from a fitted textmodel_nnseq object}
\usage{
\method{predict}{textmodel_nnseq}(
  object,
  newdata = NULL,
  type = c("class", "probability"),
  force = TRUE,
  ...
)
}
\arguments{
\item{object}{a fitted \link{textmodel_nnseq} model}

\item{newdata}{dfm on which prediction should be made}

\item{type}{the type of predicted values to be returned; see Value}

\item{force}{make \code{newdata}'s feature set conformant to the model terms}

\item{...}{not used}
}
\value{
\code{predict.textmodel_nnseq} returns either a vector of class
predictions for each row of \code{newdata} (when \code{type = "class"}), or
a document-by-class matrix of class probabilities (when \code{type =
  "probability"}).
}
\description{
\code{predict.textmodel_nnseq()} implements class predictions from a fitted
sequential neural network model.
}
\seealso{
\code{\link{textmodel_nnseq}}
}
\keyword{internal}
\keyword{textmodel}
