% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_svmlin.R
\name{predict.textmodel_svmlin}
\alias{predict.textmodel_svmlin}
\title{Prediction from a fitted textmodel_svmlin object}
\usage{
\method{predict}{textmodel_svmlin}(
  object,
  newdata = NULL,
  type = c("class", "probability"),
  force = FALSE,
  ...
)
}
\arguments{
\item{object}{a fitted linear SVM textmodel}

\item{newdata}{dfm on which prediction should be made}

\item{type}{the type of predicted values to be returned; see Value}

\item{force}{make newdata's feature set conformant to the model terms}

\item{...}{not used}
}
\value{
\code{predict.textmodel_svmlin} returns either a vector of class
predictions for each row of \code{newdata} (when \code{type = "class"}), or
a document-by-class matrix of class probabilities (when \code{type =
  "probability"}).
}
\description{
\code{predict.textmodel_svmlin()} implements class predictions from a fitted
linear SVM model.
}
\seealso{
\code{\link{textmodel_svmlin}}
}
\keyword{internal}
\keyword{textmodel}
