% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lda.R
\name{print.textmodel_lda}
\alias{print.textmodel_lda}
\title{print method for a LDA model}
\usage{
\method{print}{textmodel_lda}(x, ...)
}
\arguments{
\item{x}{for print method, the object to be printed}

\item{...}{unused}
}
\description{
print method for a LDA model
}
\keyword{internal}
\keyword{textmodel}
