% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_cnnlstmemb.R
\name{save.textmodel_cnnlstmemb}
\alias{save.textmodel_cnnlstmemb}
\title{A function for saving keras-based models}
\usage{
\method{save}{textmodel_cnnlstmemb}(x, ...)
}
\arguments{
\item{x}{model generated by \code{\link{textmodel_cnnlstmemb}}}

\item{...}{additional arguments passed to save}
}
\description{
A function for saving keras-based models
}
