% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_nnseq.R
\name{save.textmodel_nnseq}
\alias{save.textmodel_nnseq}
\title{A function for saving keras-based models}
\usage{
\method{save}{textmodel_nnseq}(x, ...)
}
\arguments{
\item{x}{model generated by \code{\link{textmodel_nnseq}}}

\item{...}{additional arguments passed to save}
}
\description{
A function for saving keras-based models
}
