% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_cnnlstmemb.R
\name{summary.textmodel_cnnlstmemb}
\alias{summary.textmodel_cnnlstmemb}
\title{summary method for textmodel_cnnlstmemb objects}
\usage{
\method{summary}{textmodel_cnnlstmemb}(object, ...)
}
\arguments{
\item{object}{output from \code{\link{textmodel_cnnlstmemb}}}

\item{...}{additional arguments not used}
}
\description{
summary method for textmodel_cnnlstmemb objects
}
\keyword{internal}
\keyword{textmodel}
