% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_nnseq.R
\name{summary.textmodel_nnseq}
\alias{summary.textmodel_nnseq}
\title{summary method for textmodel_nnseq objects}
\usage{
\method{summary}{textmodel_nnseq}(object, ...)
}
\arguments{
\item{object}{output from \code{\link{textmodel_nnseq}}}

\item{...}{additional arguments not used}
}
\description{
summary method for textmodel_nnseq objects
}
\keyword{internal}
\keyword{textmodel}
