% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_svm.R
\name{summary.textmodel_svm}
\alias{summary.textmodel_svm}
\title{summary method for textmodel_svm objects}
\usage{
\method{summary}{textmodel_svm}(object, n = 30, ...)
}
\arguments{
\item{object}{output from \code{\link[=textmodel_svm]{textmodel_svm()}}}

\item{n}{how many coefficients to print before truncating}

\item{...}{additional arguments not used}
}
\description{
summary method for textmodel_svm objects
}
\keyword{internal}
\keyword{textmodel}
