% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_svmlin.R
\name{summary.textmodel_svmlin}
\alias{summary.textmodel_svmlin}
\title{summary method for textmodel_svmlin objects}
\usage{
\method{summary}{textmodel_svmlin}(object, n = 30, ...)
}
\arguments{
\item{object}{output from \code{\link{textmodel_svmlin}}}

\item{n}{how many coefficients to print before truncating}

\item{...}{additional arguments not used}
}
\description{
summary method for textmodel_svmlin objects
}
\keyword{internal}
\keyword{textmodel}
