% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_wordfish.R
\name{summary.textmodel_wordfish}
\alias{summary.textmodel_wordfish}
\title{summary method for textmodel_wordfish}
\usage{
\method{summary}{textmodel_wordfish}(object, n = 30, ...)
}
\arguments{
\item{object}{a \link{textmodel_wordfish} object}

\item{n}{maximum number of features to print in summary}

\item{...}{unused}
}
\value{
a \code{summary.textmodel} classed list containing the call, the
estimated document positions, and the estimated feature scores
}
\description{
summary method for textmodel_wordfish
}
\keyword{internal}
\keyword{textmodel}
