% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lr.R
\name{textmodel_lr}
\alias{textmodel_lr}
\title{Logistic regression classifier for texts}
\usage{
textmodel_lr(x, y, nfolds = 10, ...)
}
\arguments{
\item{x}{the \link{dfm} on which the model will be fit.  Does not need to
contain only the training documents.}

\item{y}{vector of training labels associated with each document identified
in \code{train}.  (These will be converted to factors if not already
factors.)}

\item{nfolds}{.}

\item{...}{additional arguments passed to \code{\link[glmnet]{cv.glmnet}}}
}
\description{
Fit a fast logistic regression classifier for texts, using the \pkg{glmnet}
package.
}
\examples{
## Example from 13.1 of _An Introduction to Information Retrieval_
corp <- quanteda::corpus(c(d1 = "Chinese Beijing Chinese",
                           d2 = "Chinese Chinese Shanghai",
                           d3 = "Chinese Macao",
                           d4 = "Tokyo Japan Chinese",
                           d5 = "London England Chinese",
                           d6 = "Chinese Chinese Chinese Tokyo Japan"),
                         docvars = data.frame(train = factor(c("Y", "Y", "Y",
                                                               "N", "N", NA))))

dfmat <- quanteda::dfm(corp, tolower = FALSE)

## simulate bigger sample as classification on small samples is problematic
set.seed(1)
dfmat <- quanteda::dfm_sample(dfmat, 50, replace = TRUE)

## train model
(tmod1 <- textmodel_lr(dfmat, quanteda::docvars(dfmat, "train")))
summary(tmod1)
coef(tmod1)

## predict probability and classes
predict(tmod1, type = "prob")
predict(tmod1)
}
\seealso{
\code{\link[glmnet]{cv.glmnet}}
}
