% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.textmodels-package.R
\docType{package}
\name{textmodels}
\alias{textmodels}
\alias{_PACKAGE}
\alias{quanteda.textmodels}
\alias{quanteda.textmodels-package}
\title{quanteda.textmodels: Scaling models and classifiers for textual data}
\description{
Scaling models and classifers for sparse matrix objects representing textual data in the form of a document-feature matrix. 
    Originally from the quanteda package, these models take quanteda dfm (document-feature matrix) objects as inputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/quanteda/quanteda.textmodels}
}

}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@lse.ac.uk} (\href{https://orcid.org/0000-0002-0797-564X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kohei Watanabe \email{watanabe.kohei@gmail.com} (\href{https://orcid.org/0000-0001-6519-5265}{ORCID})
  \item Haiyan Wang \email{whyinsa@yahoo.com} (\href{https://orcid.org/0000-0003-4992-4311}{ORCID})
  \item Stefan Müller \email{mullers@tcd.ie} (\href{https://orcid.org/0000-0002-6315-4125}{ORCID})
  \item Patrick O. Perry \email{patperry@gmail.com} (\href{https://orcid.org/0000-0001-7460-127X}{ORCID})
  \item Benjamin Lauderdale \email{b.e.lauderdale@lse.ac.uk} (\href{https://orcid.org/0000-0003-3090-0969}{ORCID})
  \item William Lowe \email{wlowe@princeton.edu} (\href{https://orcid.org/0000-0002-1549-6163}{ORCID})
}

Other contributors:
\itemize{
  \item European Research Council (ERC-2011-StG 283794-QUANTESS) [funder]
}

}
\keyword{internal}
