% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{force_conformance}
\alias{force_conformance}
\title{Internal function to match a dfm features to a target set}
\usage{
force_conformance(x, features, force = TRUE)
}
\arguments{
\item{x}{input \link[quanteda:dfm]{dfm}}

\item{features}{character; a vector of feature names}

\item{force}{logical; if \code{TRUE}, make the new dfm conform to the vector of
features, otherwise return an error message}
}
\value{
a \link[quanteda:dfm]{dfm} from the \pkg{quanteda} package containing
only \code{features} as columns, in the same order as \code{features}.  A warning message
is printed if some feature names from \code{features} are not matched in \code{x}.
}
\description{
Takes a dfm and a set of features, and makes them match the features listed
in the set.
}
\examples{
quanteda.textmodels:::force_conformance(quanteda::data_dfm_lbgexample, c("C", "B", "Z"))
}
\keyword{dfm}
\keyword{internal}
