% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_cnnlstmemb.R, R/textmodel_nnseq.R
\name{save.textmodel_cnnlstmemb}
\alias{save.textmodel_cnnlstmemb}
\alias{load.textmodel_cnnlstmemb}
\alias{save.textmodel_nnseq}
\alias{load.textmodel_nnseq}
\title{Load or save keras-based textmodels}
\usage{
\method{save}{textmodel_cnnlstmemb}(x, ...)

\method{load}{textmodel_cnnlstmemb}(x, ...)

\method{save}{textmodel_nnseq}(x, ...)

\method{load}{textmodel_nnseq}(x, ...)
}
\arguments{
\item{x}{a \pkg{keras}-based fitted textmodel}

\item{...}{additional arguments passed to \code{\link[=save]{save()}} or \code{\link[=load]{load()}}}
}
\description{
Functions for loading and saving \pkg{keras}-based models.  Because these are
stored as references, they need to be "serialized" prior to saving, or
serialized upon loading.  This applies to models fit using
\code{\link[=textmodel_cnnlstmemb]{textmodel_cnnlstmemb()}} and \code{\link[=textmodel_nnseq]{textmodel_nnseq()}}.
}
\keyword{internal}
