% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lr.R
\name{summary.textmodel_lr}
\alias{summary.textmodel_lr}
\title{summary method for textmodel_lr objects}
\usage{
\method{summary}{textmodel_lr}(object, n = 30, ...)
}
\arguments{
\item{object}{output from \code{\link[=textmodel_lr]{textmodel_lr()}}}

\item{n}{how many coefficients to print before truncating}

\item{...}{additional arguments not used}
}
\value{
a \code{summary.textmodel} classed list containing elements from the
call to \code{textmodel_lr()}, including the call, statistics for lambda, and
the estimated feature scores
}
\description{
summary method for textmodel_lr objects
}
\keyword{internal}
\keyword{textmodel}
