% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lda.R
\name{textmodel_lda}
\alias{textmodel_lda}
\title{Latent Dirichlet allocation}
\usage{
textmodel_lda(
  x,
  k = 10,
  max_iter = 2000,
  seeds = NULL,
  alpha = NULL,
  beta = NULL,
  verbose = quanteda_options("verbose")
)
}
\description{
Latent Dirichlet allocation
}
\keyword{experimental}
\keyword{textmodel}
