% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_lr.R
\name{textmodel_lr}
\alias{textmodel_lr}
\title{Logistic regression classifier for texts}
\usage{
textmodel_lr(x, y, nfolds = 10, ...)
}
\arguments{
\item{x}{the \link{dfm} on which the model will be fit.  Does not need to
contain only the training documents.}

\item{y}{vector of training labels associated with each document identified
in \code{train}.  (These will be converted to factors if not already
factors.)}

\item{nfolds}{.}

\item{...}{additional arguments passed to \code{\link[glmnet]{cv.glmnet}}}
}
\description{
Fit a fast logistic regression classifier for texts, using the \pkg{glmnet}
package.
}
\examples{
\dontrun{
library("quanteda")
# use party leaders for govt and opposition classes
docvars(data_corpus_irishbudget2010, "govtopp") <-
    c(rep(NA, 4), "Govt", "Opp", NA, "Opp", NA, NA, NA, NA, NA, NA)
dfmat <- dfm(data_corpus_irishbudget2010)
tmod <- textmodel_lr(dfmat, y = docvars(dfmat, "govtopp"))
predict(tmod)
predict(tmod, type = "probability")

# multiclass problem - all party leaders
tmod2 <- textmodel_lr(dfmat,
                      y = c(rep(NA, 3), "SF", "FF", "FG", NA, "LAB",
                           NA, NA, "Green", rep(NA, 3)))
predict(tmod2)
predict(tmod2, type = "probability")
}
}
\seealso{
\code{\link[glmnet]{cv.glmnet}}
}
