% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.textmodels-package.R
\docType{package}
\name{textmodels}
\alias{textmodels}
\alias{_PACKAGE}
\alias{quanteda.textmodels}
\alias{quanteda.textmodels-package}
\title{quanteda.textmodels: Scaling Models and Classifiers for Textual Data}
\description{
Scaling models and classifiers for sparse matrix objects representing 
    textual data in the form of a document-feature matrix.  Includes original 
    implementations of Laver, Benoit, and Garry's (2003) <doi:10.1017/S0003055403000698>,
    Perry and Benoit's (2017) <arXiv:1710.08963> class affinity scaling model, and
    Slapin and Proksch's (2008) <doi:10.1111/j.1540-5907.2008.00338.x> "wordfish"
    model, as well as methods for correspondence analysis, latent semantic analysis,
    and fast Naive Bayes and linear SVMs specially designed for sparse textual data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/quanteda/quanteda.textmodels}
}

}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@lse.ac.uk} (\href{https://orcid.org/0000-0002-0797-564X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kohei Watanabe \email{watanabe.kohei@gmail.com} (\href{https://orcid.org/0000-0001-6519-5265}{ORCID})
  \item Haiyan Wang \email{whyinsa@yahoo.com} (\href{https://orcid.org/0000-0003-4992-4311}{ORCID})
  \item Stefan Müller \email{mullers@tcd.ie} (\href{https://orcid.org/0000-0002-6315-4125}{ORCID})
  \item Patrick O. Perry \email{patperry@gmail.com} (\href{https://orcid.org/0000-0001-7460-127X}{ORCID})
  \item Benjamin Lauderdale \email{b.e.lauderdale@lse.ac.uk} (\href{https://orcid.org/0000-0003-3090-0969}{ORCID})
  \item William Lowe \email{wlowe@princeton.edu} (\href{https://orcid.org/0000-0002-1549-6163}{ORCID})
}

Other contributors:
\itemize{
  \item European Research Council (ERC-2011-StG 283794-QUANTESS) [funder]
}

}
\keyword{internal}
